Program collisions
Using a lot of messy conditionals, program the board so that when Pac-Man
is about to collide with a boundary, he stops.

Program the movement the same as on the original Pac-Man where entering a
movement key will not change direction until a gap is available in the
direction the player pressed.