Test models
Create two test files for the Boundary class and PacMan class. Test the
instance variables that each have upon instantiation.

Delete the App.test.js file as it is no longer needed and the test was
incompatible with the latest code anyway.

Change the test script in package.json to now running jest and add another
one for jest coverage.