Display score on screen
Remove the console.log for the players score and add a <p> element to
the html which has its inner text changed to the current score everytime
the canvas re renders. Extract a helper method into its own file for this.

Also, in the map array, change all the empty spaces to fullstops in strings
to represent squares with pellets in.