Modify boundary collision
Add some padding to hitBoundaryConditional as well as some speed variables
in both the PacMan class and Ghost class so that it will be easy to alter
the characters speeds in future.

This will be especially important when the size of the map will change which
will alter the size of the characters.