Make PowerUp
Create the PowerUp class and its corresponding test file which is the same as
the pellet class, except the radius is bigger.

Also, modify the function changeEatenState() in the Pellet class to just
toggle between true and false without having to pass an argument in. This
is so the variable this.hasBeenEaten cannot be changed to anything other
than a boolean when using this function.