Implement powerUp effect
Add more code to the function controlling a collision between pacman and
a ghost to reset the ghosts position and add 200 points to the score when
the ghost is scared.

Now that this function deals with both scenarios, change its name to
checkPacmanGhostCollision so its name applies to any type of collision
between pacman and a ghost.