Reset ghosts after collision when scared
Add a new reset function in the Ghost class along with two new instance
variables called this.originalPosition and this.originalVelocity.

Implement reset so that when a ghost collides with pacman when scared,
the ghost gets reset back to its original position and velocity that are
specific to that instance of the Ghost class.

In order for the ghosts movement mechanics to work properly,
this.prevCollisions must also be set back to an empty array to avoid
the chance of the ghost moving through a boundary and off the board.