Add basic chomping animation
Add a simple chomping motion when PacMan is moving to the right by cutting
the arc slightly shorter and use lineTo in order to draw a line to the centre
of PacMan which will then automatically connect to the start of the arc.
This will give PacMan the appearance of a mouth.

Then, using an openRate variable, periodically add and subtract this openRate
from the angle at which the mouth is open, depending on whether the angle
is at its maximum or minimum.