Make PacMan lose lives and reset
Implement PacMan losing a life when colliding with a non scared ghost
and the game ending when all lives run out.

Create the reset function for PacMan which will be called when PacMan either
collides with non scared ghost or collects all pellets to move the next
level which will be implemented later on.