Display lives on screen
Create an object called level and using another helper function, display
this variable on the screen, indicating which level the player is up to.

Everytime the player collects all pellets on the board and moves up a
level, increase this objects number attribute by one in order to
update what is displayed on the screen.