Update board
Change the map array to match the official Pac-Man game. Update the length
of each boundary to now be 20 instead of 40 and increase the size of the
canvas.

Update the positions of the ghosts, as well as the positions and speed of
Pac-Man. Also, add in an orange ghost, like in the official game.