Make tunnel boundaries
Create a new function that adds four more boundaries to the boundaries
array. These four boundaries will be slightly off the board around the
tunnel on either side.

These boundaries must added in so that when the player or a ghost moves
through the tunnel, they cannot change direction as they get to the end
of the tunnel and start moving up the side of the board.

These four boundaries maintain the hitBoundaryConditional that dictates
movement for both PacMan and the ghosts.