Setup backend
Create an express backend server and test its use by fetching a simple
string from the backend and displaying it in a <p> element on the client
side. Add a url environment variable for heroku which will be the url that
the client will fetch from.

Delete the eslint config files as it will clash with the node_modules in
the client side.