Create Score model
Using mongoose Schema, make a new model on the server side for a players
score. Each score has a name and number of points attached to it.

Make the corresponding test file with a mongodb_helper to connect to
the test database when running the unit test. Update the scripts in
package.json to now run the unit tests from the backend as well.