Add score to database
modify endGame to make a post request to the backend to add the players
score to the database.

Create a new object called player in makeBoard that will take the name
passed into makeBoard once when it first runs and never again. This must
be done because when the name parameter from the main component is passed
along into makeBoard, it is changed into the animationId as makeBoard is
repeatedly called by itself when updating the animations on the board.