Add play again button
Allow the player to click play again button on the leaderboard page
which renders the game again and resets the game back to its original
state.

In order to do this, pass through each object in the game into the function
that deals with ending the game and resetting them once it is called.

The points have to be reset when the player wants to click play again,
as the points are required to display on the leaderboard page at the top.