Add in blank spaces for scores
If the length of the array of scores fetched from the backend is less
than 10, before setting the scores state equal to the scores fetched,
add in more score objects with hyphens for the name and points until
the length of the array is 10.

This will make a more aesthetically pleasing leaderboard if there are
few scores saved.