Replace boundaries with PNGs
Using the files obtained from Chris courses google drive, modify the
boundary class to drawImage with a new image instance variable. Change
makeBoundaries to create a new HTML image for each element in the map
and give the image a specific source, depending on which element it is.

This should replace all the blue filled squares on the board with these
images.