Add ghost sprites
Create each ghost sprite in Pexels and put them in the images folder.
Change the Ghost class to now draw a new this.image instance variable,
just like in the Boundary class. When creating the ghosts in makeGhosts.js,
pass in a starting sprite for each ghost appropraitely.

Create a new function called assignSprite, which changes the image src
for each ghost to the scaredGhostBlue sprite and call it at the beginning
of each animation frame of the ghosts.