Don't slow the ghosts down
Remove the lines in scareGhost.js that slow the ghost down due to a bug
where the ghosts will move out of bounds when they move through the tunnel.

This may be due to the ghosts position when switching between the normal
and slow speeds as the difference between them is half a pixel.