Add munch audio
Modify PacMan class to play both opening and closing munch as PacMan moves
through the board. Do this by playing each munch (which are now new instance
variables created from new Audio()), alternating between the two, whenever
this.openRate has to have its sign changed so each audio triggers when
PacMan open and closes his mouth. This syncs the audio with PacMan's chomp
animation.

Adjust the openRate value so that the chomping and munching audio play
at an appropriate rate.

This will cause Pac-Man to munch even when there are no pellets, but for
now, this will be a good start to implementing the audio.

Update the tests as necessary with two mock objects.