Make isEating
Create a new instance variable for PacMan called isEating which will be
false upon instantiation. Modify chomp to only play the munching sounds
if isEating is true.

Update the tests appropriately to now account for this new variable.

Later, functionality will be implemented for isEating to change when
Pac-Man is about to eat a pellet.