Update isEating
Create a new function that monitors whether or not PacMan is about to
collide with any pellets and changes isEating to true if he is and false
if he is not.

Now, when PacMan is eating pellets that have not been eaten, the munching
audio will play.

Update the test file for implementPacman to take in a new callback and
another object variable to mock the pellets array.