Add buttons for mobile
Add a new div element to the Game component that contains four buttons
called up, left, right and down. Only display these buttons on tablets
and phones by using the CSS @media only with a max width of 1024px.

Create an onClick function for each button that triggers a keyboard
event for each type of arrow key so that players can now play the game
on mobiles and tablets.