Fix rainbow snaked
Hard code in the board width and height in clearRect in playGame.js so
that the number of pixels its clearing is always the entire canvas.

Also adjust the width for the buttons on the leaderboard to 100% so they
don't spill over to the right side of the screen for mobile and tablets.