Apply dependency on length
Alter the code in all the object models and make functions so that they
all rely on the length parameter to define all dimensions and attributes
which relate to position and velocity. Do this so that it will be easier
to change the size of the board later, if need be.