Adjust board dimensions
Change the cell length of the board from 20 to 32. This will change the
length to be in base 2 which should stop a particular bug from appearing
with regards to the ghost movement slowing down when scared.

Update the boundary images to match the new board dimensions.

Also, fix the width of the board in CSS to fit the screen.