Slow ghosts when scared
Create a new function that checks that the ghosts speed matches their
scared state. If it doesn't, change the speed and velocity appropriately in
order to slow down the ghost when scared and speed it back up when not.

Make a new function that adjusts the position of the ghost slightly before
changing the speed and velocity of the ghost if the x or y positions are
not divisible by 4. This will need to be implemented otherwise the ghosts
could clip through boundaries due to their positions being slightly offset
by the slower speed when scared.

Also, when testing position and velocity, change the expect statement to
equal an object rather than using separate expect statements for each
dimension.