Create hunting mechanic
Make a basic hunting mechanic for the red ghost that checks which direction
is the shortest direct distance to pacman when it is about to change direction.

This won't necessarilly send the ghost on the quickest route, but is a basic
hunting movement that will be better than random movement.