Implement pink hunt
Modify chooseMovement and calculateDistance to now apply the pink ghosts
hunting movement where it aims for four spaces in front of Pac-Man. Do this
by checking Pac-Man's rotation to see which direction he is facing and
aim for the space in front.

As a side note, before I mentioned in a previous commit that the red ghosts
hunting movement was a basic version and needed to be improved. It turns out
this movement already implemented is the correct hunting movement for the
red ghost anyway so that specific movement mechanic is finished.