Fully integrate orange ghost hunt movement
Change the conditional and tests in chooseMovement to now account for the
orange ghost as well. The conditional can now be simplified to just check
if the ghost is scared as every ghost now has their hunting movement
implemented and thus the colour of the ghost in the conditional will not
matter.