Change ghosts movement while recovering
Add another conditional to when checking the speed matches the state to
adjust the ghosts speed and velocity when moving from scared to recovering
and from recovering to normal.

Add the recovering state into the conditionals of chooseMovement so that the
ghosts move randomly while recovering.

Update assignSprite to set the ghosts sprite to the eyes when recovering.

Fix a mistake on dealWithCollision so that changeRecoveringState gets called
properly in the recoveringTimeout.