Adjust position when recovering
Modify adjustPosition to correct the offset of the ghosts position when
entering recovery mode. This will need different conditionals that depend
on the direction the ghost is travelling in. This difference with the
other conditional arises due to the ghosts speed increasing and therefore,
is much easier for the ghost to clip through the wall because of the
greater distance travelled in each frame.