Create the flash rate
In PowerUp, create a new instance variable called this.rate which will be
the rate at which the radius of the power up will increase and decrease.
This change in radius will cause the power up to appear to flicker on the
board.

Creat the flash function which will add and subtract this rate over time.