Finish component testing
Create a new test to mock fetch in order to test the different outcomes
of the GET request to the backend scores.

If the fetch fails, display a message saying "Oops, something went wrong!"
onto the page.

Before the fetch has been completed, display a message in the table saying
"Please wait a moment...".