Create the variables object
Make a new object in playGame that contains all the variables listed below
it. Use the last key called start (which is a boolean) and replace the
functionality of the old start variable in the first part of playGame.

This new object will replace all the various variables below it, which
will make all the function calls shorter due to their being fewer parameters
passed into them.