Implement the visibility detection
Update the arguments passed into addVisibilityDetection in playGame so
that the full functionality of the visibility changes are implemented.

Also, pass the cycleTimer as an argument into scaredTimer.resume() in
resumeTimers so that the timeout works correctly. Update the corresponding
test to check that it is called with the correct argument.