Change the siren audio
Switch the siren audio from being attached to each ghost object to being
part of an array of audioObjects that are made by a separate function.

This should be done as it will be easier to monitor one audio object as
oppose to four. Also, having the ghost audio attached to the ghost appears
as though it actually won't have any particular advantages.