Tweak the mechanics of game over
Instead of simply running the cycle timer again in the leaderboard component,
set the start in variables back to true so that finishSetup gets called
again. Also, remove all event listeners in resetAfterGameOver so they don't
stack up whenever the player wants to play again.

This must be done so that it will be much easier to start playing the siren
audio for when the ghosts are moving in their hunt-scatter cycle.