Make shrink function
Create a new function in the PacMan class with a new constructor variable
called this.shrinkRate. Inside the new function, which will be called shrink,
descrease this.radians by this.shrinkRate.

This function will be needed so that Pac-Man will shrink into nothing when
he gets attacked by a ghost.