Update PacMan class
Modify the shrink function to add the shrinkRate because when Pac-Man shrinks,
in actual fact his mouth opens more which means this.radians needs to increase.

Also, modify the draw function to not contain the ctx calls for monitoring
the rotation of Pac-Man. This will also mean that draw can be called inside
of shrink.