Remove saveScore
Remove the call on saveScore from endGame as this will now be handled as the
death animation starts so that the score will be saved to the database
before the leaderboard component loads. This will fix this bug because the
animation will take enough time for the POST request to be completed.