Make endGame async
Change endGame to an asynchronous function with await being used saveScore
so that the players score will always be saved before the leaderboard
component starts loading. This means if the player gets a top ten score,
their score will always be displayed on the leaderboard just after their
game has ended. (As long as the POST request succeeds)