Update audio
Move the code that pauses all the ghost audio into dealWithCollision so that
it doesn't get repeatedly called in the requestAnimationFrame call in
runDeathAnimation.

Also, create the pacmanDeathAudio and pass it through the appropriate functions,
ready to be used for the animation.

Remove the cancelAnimationFrame from runDeathAnimation so that this also
doesn't get called repeatedly. This will need to be added to dealWithCollision
instead.