Unload and load the ghost audio
Call unload on each of the ghost audio objects in dealWithCollision so that
they don't overlap with the death audio.

In resetAfterDeath, load the ghost audio back in so that they play when the
game continues.

Remove the unload calls in resetAfterGameOver because they will have already
been unloaded by this point.