Alter preventDefault
Remove the arrow left and right keys from the event listener to remove the
scrolling cause by arrow keys so that the user can navigate through their
name textbox using the left and right keys. This will not affect scrolling
when playing the game unless the user zooms in which would be unnecessary
as this would stop the entire board from being visible while playing.