Change levelUp mechanic slightly
Instead of the playGame function simply carrying on when the player levels
up, cancel the animation frame and then run resetAfterLevelUp with playGame
being called at the end of resetAfterLevelUp to restart the animation frames.

This setup will allow a level up animation to be inserted in between cancelling
the animation frame and calling resetAfterLevelUp.

Also, tweak the character length requirements for the players name.