Begin basic level up animation
Make a basic piece of animation that displays the phrase "Level Up!" on the
board when the player completes a level.

Do this by cancelling the animation frame and then running a new function
called runLevelUpAnimation that adds 1 to a new variable at each step while
displaying the phrase on the board using HTML canvas fillText function. Then,
once the counter has reached 100, then resetAfterLevelUp is called (which
in turn calls playGame to continue the game).

This counter allows the time that the animation lasts on the board to be
manipulated easily.