Unload and load ghost audio
Call unload on each of the ghost audio objects just before the level up
animation begins in CheckLevelUpCondition and then call load on each of them
once the animation has finished near the end of resetAfterLevelUp.

Do this so that the ghost audio doesn't continue to play over the animation,
as there will be an audio to play for levelling up.