Play the level up audio
Call load and play on the level up audio object in checkLevelUpCondition and
then call unload in runLevelUpCondition once the animation has finished.

Also, adjust the value that the levelUpCount needs to reach to end the
animation so that there is enough time for the level up audio to fully
play.