Flash the boundaries
Add a conditional in runLevelUpAnimation that calls flash on all the
boundaries when the level up count is a multiple of 10. This will make the
boundaries appear to flash when the player is levelling up.

Also, add another condition in the conditional that does not evaluate to
true if the level up count is zero. This leaves out one round of calling
flash on the boundaries so that the boundaries are not left with the white
images after the animation has finished.