Fix server request errors
Fix the require statement in the scores controller in order to import the
correct object, change the res.render in app.js to next(err) as there is
no default engine, and bring back the index.js in the models folder in
case it is needed.

Try deleting the index.js model file to see if it is needed next.