Make new first test
Create another score testing file that uses the npm supertest package to
check that the GET request to the backend works. Add more npm scripts to
package.json for creating the database and running the migration.

Update test.yml to set up the PostgeSQL test database and run the migration
before testing both backend and frontend.