Incorporate pause detection
Create isGamePaused and pauseEventListener keys to the variables object.
Call addPauseDetection in finishSetup to add the pause event listener so that
the pause function is incorporated into the program.

(There are a couple of bugs to fix. Inconsistent time remaining values when
repeatedly pausing the timers and pausing the death animation.)