Remove event listener
In resetAfterGameOver, remove the pause event listener so that it doesn't
layer multiple event listeners of the same type where they will cancel
out when the player clicks the play again button.

Also, make a slight correction and remove the visibility event listener from
the window instead of the document.