Render game score
Call fillText on the game info ctx to display the player score above the
board. Fill the info canvas with red for dev purposes to see the text
alignment clearly. This will be deleted once finished.

Also, set the textBaseline on ctx to middle so that it is easier to align
the text in the center of the canvas nicely.

Update the dimensions of the canvas to fit the width of the div element
nicely.