Make displayPleaseWait
Create a new function that will display some text informing the player
to wait while their score is uploaded to the database. Start this function
off by adding a dark tint to the board similar to loadPauseTint so that
the text will stand out in the middle of the board.