Fix dynamic rendering
Call render on the reactRoot when clicking the Change Player button
in the Leaderboard component instead of reloading the page.

In order for createRoot to only be called once, create a conditional in
the Main component with a new reactRoot parameter that renders the Game
component if it is true (i.e. not undefined), and creates the root if not.